JDO improvements
JDO to metadata for precise schema generation.
JDO type change implications.
JAXB type changes.

+++ lib/src/main/java/hu/keve/jgc/dao/Taxtable.java
@@ -8,7 +8,7 @@ public interface Taxtable {
 
 	String getName();
 
-	short getRefcount();
+	long getRefcount();
 
 	boolean isInvisible();
